Inline AbstractBatchTest.
